Add serverless handler for Vercel deployments
