+++ go.mod
@@ -1,3 +1,3 @@
-module giscus-wrapper
+module giscus-proxy
 
 go 1.25.0
